+++ app/src/main/java/com/example/homework1/view/fragments/ItemDetailsFragment.kt
@@ -28,8 +28,8 @@ class ItemDetailsFragment : BaseFragment() {
         super.onCreate(savedInstanceState)
         viewModel = ViewModelProviders.of(activity!!).get(DetailsViewModel::class.java)
         viewModel.init()
-        binding.executePendingBindings()
     }
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
